// server/src/types/logger.ts
export interface ILog {
  level: string
  // oxlint-disable-next-line typescript/no-explicit-any
  data?: any
  message: string
  space?: boolean
}

export interface ILogs {
  logs: ILog[]
}

export interface ILogsAndErrors extends ILogs {
  errors: IError[]
}

export interface IError {
  code?: string
  title: string
  data?: string | object
  message?: string
}

export interface IException {
  name: string
  message: string
}